Выносит из utils.js функциональность экранирования в safe-html.js
