+++ src/views/card-view.js
@@ -1,5 +1,5 @@
 import View from './view.js';
-import {html} from '../utils.js';
+import {html} from '../safe-html.js';
 
 /** @extends {View<PointViewState>} */
 class CardView extends View {
