+++ src/views/editor-view.js
@@ -1,7 +1,8 @@
 import './editor-view.css';
 
 import View from './view.js';
-import {createDatePickers, html} from '../utils.js';
+import {html} from '../safe-html.js'
+import {createDatePickers} from '../utils.js';
 
 /**
  * @extends {View<PointViewState>}
